natalia392 chore: move the api request to show the puzzles for one user to 'my collection'
